Update patch to use ASM 4.x
Version of the patch is updated to 1.4 (where the .4 indicates support for ASM major version 4).
Had to add a hack and lie about the bytecode format major version to bypass an ASM 4 check
that throws IAE with Java 8 bytecode.  It seems to work, so don't complain.  Remember, we're
already patching JDK bytecode in RAM... what could go wrong?  :-)